Moved class "DrawingSelection" into DrawingSelection.js
